+++ src/crossover/__init__.py
@@ -1,1 +1,2 @@
-# From file name import function name
+# FROM (. means local) file_name IMPORT function_name
+from .examples import crossover_examples
