+++ src/crossover/examples.py
@@ -1,0 +1,3 @@
+class crossover_examples:
+    """Crossover examples will go here """
+    pass
