+++ src/fitness_function/__init__.py
@@ -1,2 +1,2 @@
-# FROM (. means local) file_name IMPORT function_name
-from .example_is_it_5 import example_is_it_5
+# FROM (. means local) file_name IMPORT class name
+from .examples import fitness_examples
