+++ src/initialization/__init__.py
@@ -1,5 +1,5 @@
 # FROM (. means local) file_name IMPORT function_name
-from .random_initialization import random_initialization
+from .examples import initialization_examples
 from .population_structure.population import population
 from .chromosome_structure.chromosome import chromosome
 from .gene_structure.gene import gene
