+++ src/mutation/__init__.py
@@ -1,0 +1,2 @@
+# FROM (. means local) file_name IMPORT function_name
+from .examples import mutation_examples
