+++ src/mutation/examples.py
@@ -1,0 +1,3 @@
+class mutation_examples:
+    """Selection examples will go here """
+    pass
