+++ src/selection/examples.py
@@ -1,0 +1,3 @@
+class selection_examples:
+    """Selection examples will go here """
+    pass
