+++ src/termination_point/__init__.py
@@ -1,3 +1,2 @@
-# FROM (. means local) file_name IMPORT function_name
-from .generation_based import generation_based
-from .fitness_based import fitness_based
+# FROM (. means local) file_name IMPORT class name
+from .examples import termination_examples
