+++ src/termination_point/examples.py
@@ -1,0 +1,16 @@
+class termination_examples:
+    """Example functions that can be used to terminate the the algorithms loop"""
+
+    def fitness_based(ga):
+        """Fitness based approach to terminate when the goal fitness has been reached"""
+        status = True
+        if(ga.current_fitness > ga.goal_fitness):
+            status = False
+        return status
+
+    def generation_based(ga):
+        """Generation based approach to terminate when the goal generation has been reached"""
+        status = True
+        if(ga.current_generation > ga.max_generations):
+            status = False
+        return status
