--- src/termination_point/fitness_based.py
@@ -1,5 +1,0 @@
-def fitness_based(ga):
-    status = True
-    if(ga.current_fitness > ga.goal_fitness):
-        status = False
-    return status
